float

+++ solar_input.py
@@ -55,13 +55,13 @@ def parse_unit_parameters(line, unit):
     **star** — объект звезды.
     """
     line_fix = line.split()
-    unit.R = int(line_fix[1])
+    unit.R = float(line_fix[1])
     unit.color = line_fix[2]
-    unit.m = int(line_fix[3])
-    unit.x = int(line_fix[4])
-    unit.y = int(line_fix[5])
-    unit.Vx = int(line_fix[6])
-    unit.Vy = int(line_fix[7])
+    unit.m = float(line_fix[3])
+    unit.x = float(line_fix[4])
+    unit.y = float(line_fix[5])
+    unit.Vx = float(line_fix[6])
+    unit.Vy = float(line_fix[7])
 
 
 def write_space_objects_data_to_file(output_filename, space_objects):
